Refactor Dropdown component and adjust CSS classes
In this commit, the Dropdown component was refactored to use interface for props, improving type checking and enhancing code readability. Meanwhile, CSS classes were adjusted in both the 'dropdown.tsx' and 'itemlist.tsx' files to better handle the website layouts and appearance. Update also includes the removal of an unnecessary comment in 'droppoints.tsx'.